Incorporated the somaticseq Pipeline.
Each component of the somaticseq pipeline was rewritten to work
similarly to other sections of the pipeline. The Somaticseq pipeline
inherits from the BasePipeline class, and the somaticseq workflow
inherits from the Workflow class. The Truthset pipeline and workflow are
now separate from the somaticseq pipeline, and should be generated
separately.